Move test fixtures and resources to own package for better separation
