+++ setup.cfg
@@ -72,7 +72,7 @@ show_source = True
 
 [tool:pytest]
 minversion = 5.4.3
-addopts = -ra -q --cov=volkswagencarnet --cov-report html
+addopts = -ra -q
 testpaths = tests
 python_files = *_test.py
