--- tests/fixtures.py
@@ -1,5 +1,0 @@
-import os
-from pathlib import Path
-
-current_path = Path(os.path.dirname(os.path.realpath(__file__)))
-resource_path = os.path.join(current_path, 'resources')
