+++ tests/fixtures/resources/dummy_cookies.pickle
